# Make detection mode run fully in background and not block IIS requests