--- a/iis/main.cpp
+++ b/iis/main.cpp
@@ -56,22 +56,10 @@ RegisterModule(DWORD, IHttpModuleRegistrationInfo* moduleInfo, IHttpServer * htt
 
         // step 3: register for server events
         //
-        auto result = moduleInfo->SetRequestNotifications(
+        return moduleInfo->SetRequestNotifications(
             factory.release(), /* module factory */
-            RQ_BEGIN_REQUEST | RQ_SEND_RESPONSE /* server event mask */,
-            RQ_END_REQUEST /* server post event mask */);
-        if (FAILED(result))
-        {
-            return result;
-        }
-
-        result = moduleInfo->SetPriorityForRequestNotification(RQ_BEGIN_REQUEST, PRIORITY_ALIAS_FIRST);
-        if (FAILED(result))
-        {
-            return result;
-        }
-
-        return moduleInfo->SetPriorityForRequestNotification(RQ_SEND_RESPONSE, PRIORITY_ALIAS_LAST);
+            RQ_BEGIN_REQUEST /* server event mask */,
+            0 /* server post event mask */);
     }
     catch (const std::bad_alloc&)
     {