--- a/iis/mymodule.h
+++ b/iis/mymodule.h
@@ -12,8 +12,10 @@
 * directly using the email address security@modsecurity.org.
 */
 
-#ifndef __MY_MODULE_H__
-#define __MY_MODULE_H__
+#pragma once
+
+#define ASIO_STANDALONE
+#include "asio/thread_pool.hpp"
 
 #include "critical_section.h"
 #include "event_logger.h"
@@ -46,8 +48,8 @@ public:
 private:
     CriticalSection cs;
     EventLogger logger;
+    asio::thread_pool threadPool;
     DWORD pageSize = 0;
     bool statusCallAlreadySent = false;
 };
 
-#endif