When ModSecurity runs in the detection-only mode, a request does not
have to be blocked awaiting to be processed and can be transmitted
immediately because the result of executing rules on it does not affect
its status.
This allows to minimize the latency overhead.

To achieve this, the processing is offloaded to a thread pool and runs
independently from the main request processing by IIS.

## Performance testing

The following table shows average latencies (ms) collected for specified requests by sending them to WAF for 1 minute using a single concurrent connection.


Request   type |  DetectionOnly,   before the fix | DetectionOnly, after the fix | Prevention, before the fix | Prevention, after the fix
-- | -- | -- | -- | --
get-none-16-fields-1kb | 89.30ms | 100.90ms | 119.44ms | 95.80ms
post-application-x-www-form-urlencoded-16-fields-1kb | 118.41ms | 117.75ms | 118.62ms | 90.94ms
post-multipart-form-data-16-fields-2kb |102.74ms | 114.99ms | 117.01ms | 91.68ms
get-none-32-fields-1kb | 121.67ms | 81.52ms | 118.21ms | 164.33ms
post-application-x-www-form-urlencoded-32-fields-1kb | 122.73ms | 85.64ms | 114.53ms | 101.09ms
post-multipart-form-data-32-fields-4kb |108.41ms | 92.98ms | 136.48ms | 115.37ms
get-none-64-fields-1kb | 106.54ms | 93.64ms | 120.33ms | 145.49ms
post-application-x-www-form-urlencoded-64-fields-1kb | 121.27ms | 76.24ms | 130.02ms | 102.79ms
post-multipart-form-data-64-fields-7kb |128.50ms | 70.86ms | 191.91ms | 104.18ms
get-none-16-fields-2kb |114.65ms | 79.81ms | 139.36ms | 114.87ms
post-application-x-www-form-urlencoded-16-fields-2kb | 59.71 |129.51ms | 98.24ms | 162.41ms | 113.76ms
post-multipart-form-data-16-fields-3kb |137.65ms | 60.45ms | 136.45ms | 97.20ms
get-none-32-fields-4kb |170.35ms | 65.85ms | 183.99ms | 134.25ms
post-application-x-www-form-urlencoded-32-fields-4kb |179.29ms | 62.49ms | 202.47ms | 89.49ms
post-multipart-form-data-32-fields-7kb |174.01ms | 44.07ms | 200.73ms | 157.27ms
post-text-plain-1-fields-1kb |80.80ms | 113.55ms | 100.36ms | 69.95ms
post-text-plain-1-fields-32kb |122.67ms | 270.86ms | 82.51ms | 83.70ms
post-text-plain-1-fields-100kb |104.03ms | 63.88ms | 123.09ms | 108.41ms
post-multipart-form-data-6-fields-1kb |93.38ms | 49.41ms | 96.41ms | 73.83ms
post-multipart-form-data-194-fields-32kb |469.74ms | 51.41ms | 506.31ms | 78.41ms
post-multipart-form-data-603-fields-100kb |1.44s | 44.86ms | 1.51s | 57.16ms
post-application-x-www-form-urlencoded-15-fields-1kb|125.46ms | 129.51ms | 121.91ms | 85.79ms
post-application-x-www-form-urlencoded-470-fields-32kb|1.16s | 183.13ms | 1.20s | 1.20s
post-application-x-www-form-urlencoded-1458-fields-100kb|3.63s | 133.29ms | 3.79s | 3.76s
post-application-json-16-fields-1kb|99.34ms | 99.58ms | 99.88ms | 88.33ms
post-application-json-512-fields-32kb|1.05s | 56.10ms | 1.10s | 236.36ms
post-application-json-1600-fields-100kb|3.30s | 60.45ms | 3.35s | 3.44s
post-text-xml-16-fields-1kb|109.28ms | 40.20ms | 96.27ms | 77.11ms
post-text-xml-512-fields-32kb|1.01s | 47.45ms | 1.01s | 966.90ms
post-text-xml-1600-fields-100kb|3.04s | 47.85ms | 3.00s | 2.98s

The following table shows average latencies (ms) collected for specified requests by sending them to WAF for 1 minute using 16 concurrent connections.

DetectionOnly,   before the fix | DetectionOnly, after the fix
-- | --
516.93ms | 83.42ms
534.47ms | 244.13ms
515.83ms | 98.01ms
652.01ms | 144.81ms
579.82ms | 100.18ms
611.76ms | 65.80ms
778.46ms | 80.35ms
770.28ms | 63.66ms
765.31ms | 75.25ms
895.39ms | 89.33ms
813.82ms | 68.21ms
809.29ms | 69.64ms
1.45s | 133.64ms
1.53s | 102.20ms
1.36s | 80.27ms
484.60ms | 196.59ms
8.46s | 300.80ms
2.77s | 221.26ms
1.80s | 214.84ms
3.84s | 168.89ms
11.98s | 281.76ms
466.62ms | 324.73ms
10.82s | 153.77ms
26.76s | 426.02ms
802.05ms | 282.12ms
9.10s | 165.82ms
26.52s | 236.57ms
562.49ms | 97.02ms
9.04s | 746.23ms
27.00s | 501.58ms
  |  

## Stress testing
Stressed the WAF setup using two simultaneously running load generators, Apache Bench and SuperBenchmarker:

```
PS C:\Tools> .\sb.exe -u "http://wafv1/?test=1=1" -c 20 -n 2000000
Starting at 5/10/2019 6:20:23 PM
[Press C to stop the test]
2000000 (RPS: 201.6)
---------------Finished!----------------
Finished at 5/10/2019 9:05:41 PM (took 02:45:18.3298147)
Status 200:    2000000

RPS: 201.6 (requests/second)
Max: 3364ms
Min: 23ms
Avg: 95.5ms

  50%   below 37ms
  60%   below 39ms
  70%   below 46ms
  80%   below 75ms
  90%   below 162ms
  95%   below 420ms
  98%   below 836ms
  99%   below 1093ms
99.9%   below 2036ms
PS C:\Tools>

./apib -x "POST" -f aaa.aaa -d 300 -c 16 "http://wafv1/?test=1=1"
Duration:             54056.659 seconds
Attempted requests:   8698267
Successful requests:  8698267
Non-200 results:      0
Connections opened:   19
Socket errors:        10

Throughput:           160.910 requests/second
Average latency:      99.430 milliseconds
Minimum latency:      24.080 milliseconds
Maximum latency:      7752.531 milliseconds
Latency std. dev:     220.323 milliseconds
50% latency:          38.614 milliseconds
90% latency:          156.977 milliseconds
98% latency:          923.683 milliseconds
99% latency:          1151.422 milliseconds

Client CPU average:    23%
Client CPU max:        49%
Client memory usage:    71%

Total bytes sent:      1352.14 megabytes
Total bytes received:  6479.22 megabytes
Send bandwidth:        0.20 megabits / second
Receive bandwidth:     0.96 megabits / second


```



All requests are responded with 200 OK, no connection resets.